# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Codecademy e-Commerce REST API",
      "description": "Defines the backend logic relating to the <strong>'registration'</strong>, <strong>'customers'</strong>, <strong>'addresses'</strong>, <strong>'orders'</strong>, <strong>'products'</strong>, <strong>'cart'</strong>, and <strong>'login'</strong> endpoints of an e-commerce REST API.  <code>GET</code> operations can be performed directly on the <strong>HEROKU</strong> database; <code>POST</code>/<code>PUT</code>/<code>DELETE</code> operations can be performed on the <strong>SWAGGER HUB</strong> <em>virtual mocking server</em> so that the <strong>HEROKU</strong> database is not compromised.",
      "version": "0.1"
    },
    "servers": [
      {
        "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
        "description": "Heroku App"
      },
      {
        "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
        "description": "SwaggerHub API Auto Mocking"
      }
    ],
    "paths": {
...
      "/api/addresses": {
        "get": {
          "tags": [
            "Addresses"
          ],
          "summary": "Return the addresses for all customers",
          "description": "Returns all the addresses in the 'addresses' table",
          "responses": {
            "200": {
              "description": "Returns a JSON array of addresses",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Address"
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "street_number": 18,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8TS",
                      "customer_id": 1,
                      "street_name": "Rosina Way"
                    },
                    {
                      "id": 2,
                      "street_number": 18,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8TS",
                      "customer_id": 2,
                      "street_name": "Rosina Way"
                    },
                    {
                      "id": 3,
                      "street_number": 18,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8TS",
                      "customer_id": 3,
                      "street_name": "Rosina Way"
                    },
                    {
                      "id": 4,
                      "street_number": 18,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8TS",
                      "customer_id": 4,
                      "street_name": "Rosina Way"
                    },
                    {
                      "id": 5,
                      "street_number": 18,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8TS",
                      "customer_id": 5,
                      "street_name": "Rosina Way"
                    },
                    {
                      "id": 6,
                      "street_number": 31,
                      "town": "Bugle",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL26 8SJ",
                      "customer_id": 10,
                      "street_name": "Springfields"
                    },
                    {
                      "id": 7,
                      "street_number": 125,
                      "town": "Upville",
                      "county": "Super Mario",
                      "country": "Oogloo",
                      "postcode": "P249 IOLUG",
                      "customer_id": 11,
                      "street_name": "Rainbow Hills"
                    },
                    {
                      "id": 8,
                      "street_number": 7,
                      "town": "Eden",
                      "county": "Paradise",
                      "country": "Omnipresence",
                      "postcode": "EVR1 WH3R3",
                      "customer_id": 12,
                      "street_name": "Day Creation"
                    },
                    {
                      "id": 9,
                      "street_number": 9,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL25 4BJ",
                      "customer_id": 6,
                      "street_name": "North Hill Park"
                    },
                    {
                      "id": 10,
                      "street_number": 9,
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL25 4BJ",
                      "customer_id": 7,
                      "street_name": "North Hill Park"
                    },
                    {
                      "id": 11,
                      "street_number": 7,
                      "town": "St. Day",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "TR16 5MR",
                      "customer_id": 8,
                      "street_name": "Trenance Road"
                    },
                    {
                      "id": 12,
                      "street_number": 10,
                      "town": "Redruth",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "TR15 1SJ",
                      "customer_id": 9,
                      "street_name": "Random Road"
                    },
                    {
                      "id": 13,
                      "street_number": 14,
                      "town": "Updated Town",
                      "county": "Updated County",
                      "country": "Updated Country",
                      "postcode": "P05t c0d3",
                      "customer_id": 13,
                      "street_name": "Updated Street Name"
                    }
                  ]
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          },
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
      "/api/addresses/{customerId}": {
        "get": {
          "tags": [
            "Addresses"
          ],
          "summary": "Return the address information for a specific customer",
          "description": "Returns the 'street_number', 'street_name', 'town', 'county', 'country', and 'postcode' fields for the customer whose id is 'customerId', in a single element array",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The customer ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful return of the relevant customer address object in a single element array",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AddressCustomer"
                    }
                  },
                  "example": [
                    {
                      "street_number": 18,
                      "street_name": "Rosina Way",
                      "town": "St. Austell",
                      "county": "Cornwall",
                      "country": "England",
                      "postcode": "PL25 8TS"
                    }
                  ]
                }
              }
            },
            "404": {
              "description": "No address found for the specified customer",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "No address is associated with this customer id."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          },
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
      "/api/addresses/{customerId}/{addressId}": {
        "put": {
          "tags": [
            "Addresses"
          ],
          "summary": "Update a particular address belonging to a specific customer",
          "description": "Updates the address that is specified by 'addressId', belonging to the customer whose id is 'customerId'",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The customer ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            },
            {
              "name": "addressId",
              "in": "path",
              "description": "The address ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressCustomer"
                },
                "examples": {
                  "0": {
                    "value": "{\n    \"street_number\": 14,\n    \"street_name\": \"Updated Street Name\",\n    \"town\": \"Updated Town\",\n    \"county\": \"Updated County\",\n    \"country\": \"Updated Country\",\n    \"postcode\": \"P05t c0d3\"\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Returns a string confirmation of successful address update",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "example": "Your address has been updated successfully."
                }
              }
            },
            "404": {
              "description": "Specified address not associated with the specified customer in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "This address is not associated with this customer id."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
...
    },
    "components": {
      "schemas": {
...
        "Address": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "The address id"
            },
            "street_number": {
              "type": "integer",
              "description": "The street number"
            },
            "street_name": {
              "type": "string",
              "description": "The street name"
            },
            "town": {
              "type": "string",
              "description": "The town"
            },
            "county": {
              "type": "string",
              "description": "The county"
            },
            "country": {
              "type": "string",
              "description": "The country"
            },
            "postcode": {
              "type": "string",
              "description": "The postcode"
            },
            "customer_id": {
              "type": "integer",
              "description": "The id of the customer who lives at this address"
            }
          }
        },
        "AddressCustomer": {
          "type": "object",
          "properties": {
            "street_number": {
              "type": "integer",
              "description": "The street number of the customer's address"
            },
            "street_name": {
              "type": "string",
              "description": "The street name of the customer's address"
            },
            "town": {
              "type": "string",
              "description": "The town associated with the customer's address"
            },
            "county": {
              "type": "string",
              "description": "The county associated with the customer's address"
            },
            "country": {
              "type": "string",
              "description": "The country associated with the customer's address"
            },
            "postcode": {
              "type": "string",
              "description": "The postcode associated with the customer's address"
            }
          }
        },
...
      }
    }
  }